+++ tests/testthat/test_releases.R
@@ -1,7 +1,7 @@
 test_that("Plotting function visualises historical
           milestones/releases of a repository", {
   testdf <- data.frame(tag_name = c("v0.1.0", "v0.1.1"),
-                       date = as.Date(c("2021-04-01", "2021-05-01")),
+                       date = c("2021-04-01", "2021-05-01"),
                        milestone = c("Minor", "Patch"))
   testplot <- plot_releases(testdf)
   expect_true(is.list(testplot))